CXXCBC-3 Move to new couchbase-cxx-client
Motivation
==========
This moves us from libcouchbase to couchbase-cxx-client. Our async
support is coming next.

Modifications
=============
* No more client wrapper on lcb.
* No more lcb.
* No more boost.
* Now with couchbase-cxx-client
* C++17 - shiny!

There is no such thing as a collection in couchbase-cxx-client - so the
API changed a bit for that -- now the bucket, scope, and collection are
all in the id of the document.

Also, why was there a collection pointer in `attempt_context.replace` or
`attempt_context.remove`?   The `transaction_get_result` has all that
information already.   So, removed it.

Added some transaction tests, removed all the client tests (and the
client itself).

Made the example game_server.cxx multithreaded, since it can then be a
bit of a stress-test, exercising the rollback pathways since there will
be tons of write-write conflicts possible.

Will need to update the jenkins scripts so we can package things, etc...
Coming soon...

I didn't correct the api documentation.  Will do later as we flesh out
the async stuff.

Results
=======
Tests pass.

Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-cxx/+/161835
Tested-by: Build Bot <build@couchbase.com>

+++ src/transactions/atr_ids.hxx
@@ -25,8 +25,8 @@ namespace transactions
     class atr_ids
     {
       public:
-        static const std::string& atr_id_for_vbucket(int vbucket_id);
-        static int vbucket_for_key(const std::string& key);
+        static const std::string& atr_id_for_vbucket(size_t vbucket_id);
+        static size_t vbucket_for_key(const std::string& key);
         static const std::vector<std::string>& all();
     };
